refactor(membudget): extract memory-budget detection into internal/membudget
Move detectMemoryBudget/readCgroupBytes/readMemTotal (cgroup v2 -> v1 ->
/proc/meminfo) out of the root-package memlimit.go into a reusable
internal/membudget package so internal/storage can budget-scale the SQLite
PRAGMA stanza without importing package main. applyMemoryLimit stays as a
thin wrapper with unchanged behavior; detection tests moved across verbatim
and the wrapper gained its own tests (operator GOMEMLIMIT honored, budget
fraction applied).
